Products error handling

+++ controllers/Products.js
@@ -10,45 +10,56 @@ export default {
     getProducts : async (req, res, next) => {
         try {
             const products = await getProducts();
-            res.status(200).json(products);
+            if (products.length > 0)
+                return res.status(200).json(products);
+            return res.status(404).json({ "message": "Products not found" });
         } catch (err) {
-            res.status(500).json({"message": `Error while getting products. Err: ${err}`});
+            console.error(`Error while getting products: ${err}`);
+            return res.status(500).json({"message": `Error while getting products. Err: ${err}`});
         }
     },
     getProductById : async (req, res, next) => {
         try {
             const product = await getProductById(req.params.id);
-            res.status(200).json(product);
+            if (product)
+                return res.status(200).json(product);
+            return res.status(404).json({ "message": "Product not found" });
         } catch (err) {
-            res.status(500).json({"message": `Error while getting product. Err: ${err}`});
+            console.error(`Error while getting product: ${err}`);
+            return res.status(500).json({"message": `Error while getting product. Err: ${err}`});
         }
     },
     addProduct : async (req, res, next) => {
         try {
             const product = await addProduct(req.body);
-            res.status(201).json(product);
+            return res.status(201).json(product);
         } catch (err) {
-            res.status(500).json({"message": `Error while adding product. Err: ${err}`});
+            console.error(`Error while adding product: ${err}`);
+            return res.status(500).json({"message": `Error while adding product. Err: ${err}`});
         }
     },
     updateProduct : async (req, res, next) => {
         try {
-            const product = await updateProduct(req);
-            res.status(202).json(product);
+            const product = await getProductById(req.params.id);
+            if (!product)
+                return res.status(404).json({ "message": "Product not found" });
+            const updatedProduct = await updateProduct(product, req.body);
+            return res.status(202).json(updatedProduct);
         } catch (err) {
-            res.status(500).json({"message": `Error while updating product. Err: ${err}`});
+            console.error(`Error while adding product: ${err}`);
+            return res.status(500).json({"message": `Error while updating product. Err: ${err}`});
         }
     },
     deleteProduct : async (req, res, next) => {
         try {
-            const success = await deleteProduct(req.params.id)
-            if (success) {
-                res.status(200).json({ "message": "Product deleted successfully" });
-            } else {
-                res.status(404).json({ "message": "Product not found" });
-            }
+            const product = await getProductById(req.params.id);
+            if (!product)
+                return res.status(404).json({ "message": "Product not found" });
+            await deleteProduct(req.params.id)
+            return res.status(200).json({ "message": "Product deleted successfully" });
         } catch (err) {
-            res.status(500).json({"message": `Error while deleting product. Err: ${err}`});
+            console.error(`Error while deleting product: ${err}`);
+            return res.status(500).json({"message": `Error while deleting product. Err: ${err}`});
         }
     }
 };
